requirements.txt
django 
djangorestframework
couldinary
djangorestframework-simplejwt
